refactor: move InstallerDialog to diff pkg
